backend/app/requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
pydantic==1.10.9
asyncpg==0.27.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pyjwt
python-multipart